sponnect/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-Caching==2.1.0
SQLAlchemy==2.0.23
celery==5.3.4
redis==5.0.1
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
pytz==2023.3.post1
email-validator==2.1.0
requests==2.31.0

